Move FastAPI server-related functions to service-specific utilities
