+++ caterva2/services/srv_utils.py
@@ -13,6 +13,7 @@ import safer
 
 # Requirements
 import blosc2
+import fastapi
 import httpx
 import tqdm
 import numpy as np
@@ -221,6 +222,30 @@ def download(host, dataset, params, localpath=None, verbose=False):
 
     return array, schunk
 
+
+#
+# HTTP server helpers
+#
+def raise_bad_request(detail):
+    raise fastapi.HTTPException(status_code=400, detail=detail)
+
+
+def raise_not_found(detail='Not Found'):
+    raise fastapi.HTTPException(status_code=404, detail=detail)
+
+
+def get_abspath(root, path):
+    abspath = root / path
+
+    # Security check
+    if root not in abspath.parents:
+        raise_bad_request(f'Invalid path {path}')
+
+    # Existence check
+    if not abspath.is_file():
+        raise_not_found()
+
+    return abspath
 
 
 #
